Microoptimisations (#6)
* gfx: Use simpler equality test for IPPU.FrameCount

As it's only incremented here and then reset to 0, avoids integer
division, which can be be expensive on CPUs without this instruction
such as pre-ARMv7 classic ARMs.

* dsp1: Use clz CPU instruction when normalising fixed-point numbers

Use the gcc built-in function, making the operation constant-time and
branch-free.

* fx: Optimise copying registers on LE platforms

Use {READ,WRITE}_WORD so that it copies by word on little-endian
platforms. Explicitly define FAST_LSB_WORD_ACCESS on Miyoo platform as
word accesses in this compilation unit are aligned.

Mark LUTs const for const-correctness.

* dsp: Fix assignment of Op11Xr

Fix regression introduced in snes9x-1.40.